HsLib

Description

Haskell equivalent functions of the CPP functions to test for correctness.

Documentation

square :: Num a => a -> a

Square a number.

isTriple :: Int -> Int -> Int -> Bool

Test whether a, b and c form a Pythagorean triple given that a < b < c. For example,

>>> isTriple 7 24 25
True
>>> isTriple 7 24 26
False

triples :: [(Int, Int, Int)]

List of Pythagorean triples for c < 100. Used to ensure that 20% of the random triple values generated are valid Pythagorean triples.

applyXorCipher :: ByteString -> ByteString -> ByteString

Used to encode or decode a message with XOR and key. For example,

>>> import qualified Data.ByteString.Char8 as C
>>> applyXorCipher (C.pack "message") (C.pack "my key")
"\NUL\FSS\CAN\EOT\RS\b"
>>> applyXorCipher (C.pack "\NUL\FSS\CAN\EOT\RS\b") (C.pack "my key")
"message"