InlineCPP

Description

Haskell wrapper functions for using CPP directly as inline source code.

Documentation

square :: CInt -> IO CInt

Inline call to CPP function cn::square. Equivalent to HS function square

isTriple :: CInt -> CInt -> CInt -> IO CInt

Inline call to CPP function cn::isTriple. Equivalent to HS function isTriple.

sumVec :: [CInt] -> IO CInt

Inline CPP that sums items in a HS list. For example,

>>> sumVec [10,20,30]
100

rangeList :: Int -> IO [Int]

Inline CPP that creates a HS list containing [0 .. (n - 1)]. For example,

>>> rangeList 4
[0,1,2,3]

testApplyXorCipher :: IO ()

Inline CPP calls cn::applyXorCipher twice to encoded then decode a message. Demonstrates use of cout and successful round trip use of the XOR cipher.

applyXorCipher :: String -> String -> IO String

Inline call to CPP function cn::applyXorCipher. Equivalent to HS function applyXorCipher. Used to encode or decode a message with XOR and key. For example,

>>> applyXorCipher "message" "my key"
"\NUL\FSS\CAN\EOT\RS\b"
>>> applyXorCipher "\NUL\FSS\CAN\EOT\RS\b" "my key"
"message"